requirements.txt
django
whitenoise
djangorestframework
pillow
djangorestframework-simplejwt